requirements.txt
fastapi
uvicorn[standard]
pydantic>=2
pydantic-settings
python-dotenv
structlog
httpx
openai>=1.40.0
tiktoken
ujson
streamlit
requests


